Added GLTexture prototype
	modified:   ../gfx/ImageAllocatorGL.cpp
	new file:   GLTexture.cpp
	new file:   GLTexture.h
	modified:   ../gui/GFXGL.cpp
	modified:   ../gui/internal/ApplicationX11.cpp

git-svn-id: https://svnknoll.informatik.tu-muenchen.de/kloses/visionTools@461 f7bcf5df-4c73-0410-a0d8-b61c104701fe

+++ cvt/gui/GFXGL.cpp
@@ -145,8 +145,8 @@ namespace cvt {
 		if( img.memType() != IALLOCATOR_GL )
 			return;
 
-		w = img.width();
-		h = img.height();
+		w = ( GLint ) img.width();
+		h = ( GLint ) img.height();
 		GLint vertices[ 8 ] = {
 			x	 , y + h,
 			x	 , y    ,
